Add support for xxhash checksums
The on disk format of the sstable footer has changed to include a
version field which can be used to determine the footer size and the
syntax/semantics of its fields.  The plain and block based table magic
numbers have been updated to bootstrap this change.  The prior format is
considered legacy and is denoted version 0.  Footer::DecodeFrom()
contains a code path to support the legacy footer format (using the
prior magic numbers for detection since it did not include the footer
version).

Footer format version 1 contains the following fields:
 - checksum type
 - metaindex handle
 - index handle
 - footer version (32 bits)
 - table magic number (64 bits)

The checksum type is a new option which can be used to select between
different schemes, currently CRC32C (default) and xxhash are supported.